PYIC-3472: Move common-services domain files to gpg45.domain
