+++ libs/gpg45-evaluator/src/main/java/uk/gov/di/ipv/core/library/gpg45/validators/Gpg45DcmawValidator.java
@@ -1,7 +1,7 @@
 package uk.gov.di.ipv.core.library.gpg45.validators;
 
 import uk.gov.di.ipv.core.library.annotations.ExcludeFromGeneratedCoverageReport;
-import uk.gov.di.ipv.core.library.domain.CredentialEvidenceItem;
+import uk.gov.di.ipv.core.library.gpg45.domain.CredentialEvidenceItem;
 
 public class Gpg45DcmawValidator {
     @ExcludeFromGeneratedCoverageReport
