+++ libs/gpg45-evaluator/src/test/java/uk/gov/di/ipv/core/library/gpg45/validators/Gpg45EvidenceValidatorTest.java
@@ -1,7 +1,7 @@
 package uk.gov.di.ipv.core.library.gpg45.validators;
 
 import org.junit.jupiter.api.Test;
-import uk.gov.di.ipv.core.library.domain.CredentialEvidenceItem;
+import uk.gov.di.ipv.core.library.gpg45.domain.CredentialEvidenceItem;
 
 import java.util.Collections;
